content/canvas/test/test_2d.state.saverestore.strokeStyle.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.state.saverestore.strokeStyle</title>
<!-- Testing: save()/restore() works for strokeStyle -->
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

// Test that restore() undoes any modifications
var old = ctx.strokeStyle;
ctx.save();
ctx.strokeStyle = "#ff0000";
ctx.restore();
ok(ctx.strokeStyle === old, "ctx.strokeStyle === old");

// Also test that save() doesn't modify the values
ctx.strokeStyle = "#ff0000";
old = ctx.strokeStyle;
    // we're not interested in failures caused by get(set(x)) != x (e.g.
    // from rounding), so compare against d instead of against "#ff0000"
ctx.save();
ok(ctx.strokeStyle === old, "ctx.strokeStyle === old");
ctx.restore();

SimpleTest.finish();

});
</script>
